`tensorflow`: add `tensorflow.autodiff` (#11442)
